improve the test suite
* The test suite now relies on smaller and cleaner functions.
* Added some new tests, mostly to check the allocation functions.
* data/input/EGM96-degree10-tbl.txt, data/input/EGM96-degree10.gfc:
  modified to match data/input/EGM96-degree10-mtx.txt.

+++ tests/cmp_arrays.c
@@ -3,8 +3,6 @@
 #include <config.h>
 #include <stdio.h>
 #include "../src/prec.h"
-#include "../src/misc/misc_is_nearly_equal.h"
-#include "parameters.h"
 #include "cmp_vals.h"
 /* ------------------------------------------------------------------------- */
 
@@ -15,9 +13,9 @@
 
 /* Compares elements of two arrays up to some threshold */
 /* ------------------------------------------------------------------------- */
-int cmp_arrays(REAL *arr1, REAL *arr2, size_t n, REAL eps)
+long int cmp_arrays(REAL *arr1, REAL *arr2, size_t n, REAL eps)
 {
-    int ret = 0;
+    long int ret = 0;
 
 
     for (size_t i = 0; i < n; i++)
